Fleshed out the components and the passing of data through them. Not supper hyped about how I had to move files around in the src folder. May ddress this later
