--- instagram-app/components/SearchBar/SearchBar.js
@@ -1,19 +1,0 @@
-import React from 'react';
-import './SearchBar.css';
-
-class SearchBar extends React.Component{
-    constructor(props){
-        super(props   );
-        this.state = {
-
-        }
-    }
-
-    render(){
-        return(
-            <h1>This is the Search Bar</h1>
-        )
-    }
-}
-
-export default SearchBar;
