+++ instagram-app/src/App.js
@@ -1,25 +1,18 @@
 import React from 'react';
 import './App.css';
+import dummyData from './dummy-data.js'
 
-import CommentSection from './components/CommentSection/CommentSection.js'
 import PostContainer from './components/PostContainer/PostContainer.js';
-import SearBar from './components/SearchBar/SearchBar.js';
+import SearchBar from './components/SearchBar/SearchBar.js';
+
+const data = dummyData;
 
 class App extends React.Component {
-
-  constructor(){
-    super();
-    this.state = {
-      
-    }
-  }
-
   render(){
     return (
       <div className="App">
-        <SearBar />
-        <PostContainer />
-        <CommentSection />
+        <SearchBar />
+        <PostContainer data = {data} />
       </div>
     );
   }
