+++ instagram-app/src/components/PostContainer/PostContainer.js
@@ -1,17 +1,21 @@
 import React from 'react';
 import './PostContainer.css';
+import Post from './Post.js'
 
 class PostContainer extends React.Component{
     constructor(props){
         super(props);
         this.state = {
-
+            data: props.data
         }
     }
 
     render(){
         return (
-            <h2>This is the Post Container</h2>
+            <div>{this.state.data.map(post =>(
+                    <Post data={post} />
+            ))}
+            </div>
         )
     }
 }
