+++ instagram-app/src/components/SearchBar/assets/heart_icon.png
