+++ instagram-app/src/dummy-data.js
@@ -1,99 +1,83 @@
 const dummyData = [
-    {
-        id: "a",
+  {
+    username: "philzcoffee",
+    thumbnailUrl:
+      "https://tk-assets.lambdaschool.com/ecd33d34-c124-4b75-92d2-e5c52c171ed8_11201517_887808411287357_1307163552_a.jpg",
+
+    imageUrl:
+      "https://tk-assets.lambdaschool.com/69cf901b-f96d-466e-a745-ff2a01effac9_philz-image.jpg",
+    likes: 400,
+    timestamp: "July 17th 2017, 12:42:40 pm",
+    comments: [
+      {
         username: "philzcoffee",
-        thumbnailUrl: 'https://instagram.fbna1-1.fna.fbcdn.net/vp/31c45306e6ca815b8c59940aed10cedc/5B2D3C88/t51.2885-19/11201517_887808411287357_1307163552_a.jpg',
-        imageUrl: 'https://instagram.fbna1-1.fna.fbcdn.net/vp/557a29d3526bc7f813c2a784d82de7d3/5B4E3E05/t51.2885-15/e35/26065916_2007800899433580_5879937492399947776_n.jpg',
-        likes: 400,
-        timestamp: "July 17th 2017, 12:42:40 pm",
-        comments: [
-            {
-                id: "1",
-                username: "philzcoffee",
-                text: "We've got more than just delicious coffees to offer at our shops!"
-            },
-            {
-                id: "2",
-                username: "biancasaurus",
-                text: "Looks delicious!"
-            },
-            {
-                id: "3",
-                username: "martinseludo",
-                text: "Can't wait to try it!"
-            }
-        ]
-    },
-    {
-        id: "b",
+        text:
+          "We've got more than just delicious coffees to offer at our shops!"
+      },
+      {
+        username: "biancasaurus",
+        text: "Looks delicious!"
+      },
+      {
+        username: "martinseludo",
+        text: "Can't wait to try it!"
+      }
+    ]
+  },
+  {
+    username: "fortnite",
+    thumbnailUrl:
+      "https://tk-assets.lambdaschool.com/ce601fdf-7cb0-4098-83d3-1a1584a72513_30841289_342445456281079_112845064497004544_n.jpg",
+    imageUrl:
+      "https://tk-assets.lambdaschool.com/89d13918-b7a2-4b40-9658-f376ea3f6b59_37131538_213683546146400_1083714364399157248_n.jpg",
+    likes: 4307,
+    timestamp: "July 15th 2017, 03:12:09 pm",
+    comments: [
+      {
         username: "twitch",
-        thumbnailUrl: 'https://instagram.fbna1-1.fna.fbcdn.net/vp/81f608ddb76a034e06c0d1669e7e2b48/5B35C840/t51.2885-19/10684037_1512272482349821_444823433_a.jpg',
-        imageUrl: 'https://instagram.fbna1-1.fna.fbcdn.net/vp/830c01ad7ceb325547fd1d4d1a1e6897/5B465520/t51.2885-15/e35/25006313_192089151354555_3374696275667582976_n.jpg',
-        likes: 4307,
-        timestamp: "July 15th 2017, 03:12:09 pm",
-        comments: [
-            {
-                id: "4",
-                username: "twitch",
-                text: "Epic Street Fighter action here in Las Vegas at #EVO2017!"
-            },
-            {
-                id: "5",
-                username: "michaelmarzetta",
-                text: "Omg that match was crazy"
-            },
-            {
-                id: "6",
-                username: "themexican_leprechaun",
-                text: "What a setup"
-            },
-            {
-                id: "7",
-                username: "dennis_futbol",
-                text: "It that injustice"
-            },
-            {
-                id: "8",
-                username: "dennis_futbol",
-                text: "Is"
-            }
-        ]
-    },
-    {
-        id: "c",
+        text: "Epic Street Fighter action here in Las Vegas at #EVO2017!"
+      },
+      {
+        username: "michaelmarzetta",
+        text: "Omg that match was crazy"
+      },
+      {
+        username: "themexican_leprechaun",
+        text: "What a setup"
+      },
+      {
+        username: "dennis_futbol",
+        text: "It that injustice"
+      },
+      {
+        username: "dennis_futbol",
+        text: "Is"
+      }
+    ]
+  },
+  {
+    username: "playhearthstone",
+    thumbnailUrl:
+      "https://tk-assets.lambdaschool.com/c432f179-8bd7-4758-959d-d88a21f96bca_37105899_432228420518610_5035444036064772096_n.jpg",
+    imageUrl:
+      "https://tk-assets.lambdaschool.com/43bf01f9-319c-469d-8cf5-0120fe1007f1_yosemite.jpg",
+    likes: 5306,
+    timestamp: "July 14th 2017, 10:04:08 am",
+    comments: [
+      {
         username: "playhearthstone",
-        thumbnailUrl: 'https://instagram.fbna1-1.fna.fbcdn.net/vp/51d5b37438ae3a47df37b7ed3fda141f/5B4ABAA7/t51.2885-19/s150x150/13398400_140638743023210_310840336_a.jpg',
-        imageUrl: 'https://instagram.fbna1-1.fna.fbcdn.net/vp/22618be7ee2a2b676d13e8d70d7d5e08/5B40BF6B/t51.2885-15/e35/25038917_1978298569058775_6081161469041311744_n.jpg',
-        likes: 5306,
-        timestamp: "July 14th 2017, 10:04:08 am",
-        comments: [
-            {
-                id: "9",
-                username: "playhearthstone",
-                text: "Power alone is not to be feared. Fear instead those who wield it! #FrozenThrone #Expansion #DeathKnights"
-            },
-            {
-                id: "10",
-                username: "tapmelon",
-                text: "Wish that death knight could be added as a new playable class in this expansion."
-            },
-            {
-                id: "11",
-                username: "micpetboudreau",
-                text: "Can't wait"
-            },
-            {
-                id: "12",
-                username: "awaywetravel",
-                text: "I <3 Hearthstone."
-            },
-            {
-                id: "13",
-                username: "awesomebt28",
-                text: "I like how gul'dan looks so old and useless"
-            }
-        ]
-    }
+        text: "Love this shot!"
+      },
+      {
+        username: "awaywetravel",
+        text: "Yosemite is my most favorite place in the universe"
+      },
+      {
+        username: "awesomebt28",
+        text: "I like how Half Dome looks so old and useless"
+      }
+    ]
+  }
 ];
 
 export default dummyData;
